# babson-scheduler/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Babson Library Student Worker Scheduler",
    "description": "\n    Constraint-aware scheduling system for Babson Library student workers.\n\n    Uses Google OR-Tools CP-SAT for optimization-based schedule generation.\n    Designed for IBM Orchestrate integration via OpenAPI custom extension.\n    ",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    },
    {
      "url": "https://your-deployment-host.com",
      "description": "Production"
    }
  ],
  "paths": {
    "/api/v1/students/": {
      "get": {
        "tags": [
          "Students"
        ],
        "summary": "List Students",
        "operationId": "list_students_api_v1_students__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StudentOut"
                  },
                  "type": "array",
                  "title": "Response List Students Api V1 Students  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Students"
        ],
        "summary": "Create Student",
        "operationId": "create_student_api_v1_students__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/students/{student_id}": {
      "get": {
        "tags": [
          "Students"
        ],
        "summary": "Get Student",
        "operationId": "get_student_api_v1_students__student_id__get",
        "parameters": [
          {
            "name": "student_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Student Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Students"
        ],
        "summary": "Update Student",
        "operationId": "update_student_api_v1_students__student_id__patch",
        "parameters": [
          {
            "name": "student_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Student Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Students"
        ],
        "summary": "Deactivate Student",
        "operationId": "deactivate_student_api_v1_students__student_id__delete",
        "parameters": [
          {
            "name": "student_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Student Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Deactivate Student Api V1 Students  Student Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/shifts/templates": {
      "get": {
        "tags": [
          "Shifts"
        ],
        "summary": "List Templates",
        "operationId": "list_templates_api_v1_shifts_templates_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ShiftTemplateOut"
                  },
                  "type": "array",
                  "title": "Response List Templates Api V1 Shifts Templates Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Shifts"
        ],
        "summary": "Create Template",
        "operationId": "create_template_api_v1_shifts_templates_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShiftTemplateCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShiftTemplateOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/shifts/templates/{template_id}": {
      "delete": {
        "tags": [
          "Shifts"
        ],
        "summary": "Deactivate Template",
        "operationId": "deactivate_template_api_v1_shifts_templates__template_id__delete",
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Template Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Deactivate Template Api V1 Shifts Templates  Template Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/shifts/instances/generate": {
      "post": {
        "tags": [
          "Shifts"
        ],
        "summary": "Generate Instances",
        "operationId": "generate_instances_api_v1_shifts_instances_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShiftInstanceGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Generate Instances Api V1 Shifts Instances Generate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/shifts/instances": {
      "get": {
        "tags": [
          "Shifts"
        ],
        "summary": "List Instances",
        "operationId": "list_instances_api_v1_shifts_instances_get",
        "parameters": [
          {
            "name": "week_start_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Week Start Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ShiftInstanceOut"
                  },
                  "title": "Response List Instances Api V1 Shifts Instances Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ShiftInstanceGenerateRequest": {
        "properties": {
          "week_start_date": {
            "type": "string",
            "format": "date",
            "title": "Week Start Date",
            "description": "Monday of the target week"
          },
          "is_exam_period": {
            "type": "boolean",
            "title": "Is Exam Period",
            "description": "Use 24/7 exam-period hours",
            "default": false
          },
          "slots_per_window": {
            "type": "integer",
            "maximum": 4.0,
            "minimum": 1.0,
            "title": "Slots Per Window",
            "description": "Concurrent desk positions per time window (default 2)",
            "default": 2
          },
          "force": {
            "type": "boolean",
            "title": "Force",
            "description": "Delete and regenerate if instances already exist",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "week_start_date"
        ],
        "title": "ShiftInstanceGenerateRequest"
      },
      "ShiftInstanceOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "template_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Id"
          },
          "week_start_date": {
            "type": "string",
            "title": "Week Start Date"
          },
          "date": {
            "type": "string",
            "title": "Date"
          },
          "day_of_week": {
            "type": "integer",
            "title": "Day Of Week"
          },
          "start_time": {
            "type": "string",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "title": "End Time"
          },
          "duration_hours": {
            "type": "number",
            "title": "Duration Hours"
          },
          "is_hard_shift": {
            "type": "boolean",
            "title": "Is Hard Shift"
          },
          "is_exam_period": {
            "type": "boolean",
            "title": "Is Exam Period"
          },
          "slot_index": {
            "type": "integer",
            "title": "Slot Index"
          }
        },
        "type": "object",
        "required": [
          "id",
          "template_id",
          "week_start_date",
          "date",
          "day_of_week",
          "start_time",
          "end_time",
          "duration_hours",
          "is_hard_shift",
          "is_exam_period",
          "slot_index"
        ],
        "title": "ShiftInstanceOut"
      },
      "ShiftTemplateCreate": {
        "properties": {
          "day_of_week": {
            "type": "integer",
            "maximum": 6.0,
            "minimum": 0.0,
            "title": "Day Of Week",
            "description": "0=Monday \u2026 6=Sunday"
          },
          "start_time": {
            "type": "string",
            "title": "Start Time",
            "description": "HH:MM"
          },
          "end_time": {
            "type": "string",
            "title": "End Time",
            "description": "HH:MM (use 00:00\u201302:00 for cross-midnight)"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label",
            "description": "Human label e.g. 'Late Night', 'Opening'"
          },
          "is_hard_shift": {
            "type": "boolean",
            "title": "Is Hard Shift",
            "description": "Override hard-shift flag",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "day_of_week",
          "start_time",
          "end_time"
        ],
        "title": "ShiftTemplateCreate"
      },
      "ShiftTemplateOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "day_of_week": {
            "type": "integer",
            "title": "Day Of Week"
          },
          "start_time": {
            "type": "string",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "title": "End Time"
          },
          "duration_hours": {
            "type": "number",
            "title": "Duration Hours"
          },
          "is_hard_shift": {
            "type": "boolean",
            "title": "Is Hard Shift"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "created_by": {
            "type": "string",
            "title": "Created By"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "day_of_week",
          "start_time",
          "end_time",
          "duration_hours",
          "is_hard_shift",
          "label",
          "created_by",
          "is_active"
        ],
        "title": "ShiftTemplateOut"
      },
      "StudentCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Full name"
          },
          "email": {
            "type": "string",
            "title": "Email",
            "description": "Unique institutional email"
          },
          "seniority_date": {
            "type": "string",
            "format": "date",
            "title": "Seniority Date",
            "description": "Employment start date (YYYY-MM-DD). Earlier = higher seniority."
          },
          "min_hours": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 8.0,
            "title": "Min Hours",
            "description": "Hard weekly hours floor (\u22658)",
            "default": 8
          },
          "max_hours": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 8.0,
            "title": "Max Hours",
            "description": "Hard weekly hours ceiling (\u226420)",
            "default": 20
          },
          "target_hours": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 8.0,
            "title": "Target Hours",
            "description": "Student's preferred weekly hours target (midpoint of their stated band)",
            "default": 8
          }
        },
        "type": "object",
        "required": [
          "name",
          "email",
          "seniority_date"
        ],
        "title": "StudentCreate"
      },
      "StudentOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "seniority_date": {
            "type": "string",
            "title": "Seniority Date"
          },
          "min_hours": {
            "type": "integer",
            "title": "Min Hours"
          },
          "max_hours": {
            "type": "integer",
            "title": "Max Hours"
          },
          "target_hours": {
            "type": "integer",
            "title": "Target Hours"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "email",
          "seniority_date",
          "min_hours",
          "max_hours",
          "target_hours",
          "is_active",
          "created_at"
        ],
        "title": "StudentOut"
      },
...
      "StudentUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "seniority_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seniority Date"
          },
          "min_hours": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 20.0,
                "minimum": 8.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Hours"
          },
          "max_hours": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 20.0,
                "minimum": 8.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Hours"
          },
          "target_hours": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 20.0,
                "minimum": 8.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Hours"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "StudentUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}